Elaborate on catching and rethrowing events' API. - add autocatch($) prefix for auto-rethrowing conditions - add call_again() for re-invoking a cancelled condition within catch{}; - add is_cancelling which is only true inside catch {}
WATCH_CANCEL is now called with min. 2 parameters,
the 2nd is the record being cancelled. It is also nulled
right befire the call to avoid endless recursion. Extended f
form of bind() is updated accordingly; Mutex uses that form
and is updated too.

HTTP uses autocatch to deposit a connection back gracefully
even if lambda is cancelled